renaming the subscription screen
